feat(logic) : criando logica do diario de aula + hora aula e ajustando funcoes em turmas e usuarios
